* support/regression/tests/bug1757671.c: fixed for 64 bit arch
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@5921 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ support/regression/tests/bug1757671.c
@@ -3,6 +3,7 @@
  */
 
 #include <testfwk.h>
+#include <stdint.h>
 
 struct c
 {
@@ -18,7 +19,7 @@ void clamp(struct c *l)
 }
 
 //char cmplong(char x, char y)
-char cmplong(long x, long y)
+char cmplong(int32_t x, int32_t y)
 {
 	if (x < y)
 		return 1;
@@ -30,9 +31,9 @@ void testBug(void)
 //	volatile char x = 0xBF;//800000; //-1.0
 //	volatile char y = 0x3F;//800000; //+1.0
 //	volatile char z = 0x43;//700000; //+240.0
-	volatile long x = 0xBF800000; //-1.0
-	volatile long y = 0x3F800000; //+1.0
-	volatile long z = 0x43700000; //+240.0
+	volatile int32_t x = 0xBF800000; //-1.0
+	volatile int32_t y = 0x3F800000; //+1.0
+	volatile int32_t z = 0x43700000; //+240.0
 	struct c val = { -1.0 };
 	clamp(&val);
 	ASSERT (val.r == 1.0);
